visp_mbox : Clear rpu mbox observation register after  fifo read.
The mbox driver uses the item_stored variable to keep track of fifo.
APU updates this shared variable after reading from the fifo and RPU updates this
after writing to fifo.

When the mbox observation register is cleared by the APU before completing
the fifo copy, it creates a race-condition where both APU & RPU update
the item_stored variable simultaneously. This leads to an invalid item_stored count.

To avoid this, the observation register is cleared by the APU after completing
fifo memcpy. This ensures that only one core access the item_stored at a
time.

Signed-off-by: Ranjith K <RanjithKumar.Katravula@amd.com>
Signed-off-by: Anil Kumar Mamidala <anil.mamidala@amd.com>