isp-modules: visp_mbox: implement shutdown callback to support kexec
The visp_mbox driver now implements the `shutdown()` callback to
ensure proper cleanup and quiescence of the mailbox device during
system shutdown or reboot.

The shutdown implementation performs the following actions:
- Frees the mailbox TX and RX channels using `mbox_free_channel()`.
- Cleans up reserved memory regions by calling
  `visp_mbox_reserved_memory_exit()`.
- Calls `visp_mbox_rpu_remove()` to clean up internal RPU structures.

Typically, kexec is used for booting a new kernel image directly from a
running kernel, bypassing the bootloader stage. The shutdown () callback
is generally used to cleanly quiesce hardware during system shutdown or
reboot to prevent data corruption.

However, since our driver is a module driver, the shutdown callback is not
executed in practice. Moreover, the .remove() callback is sufficient to
clean up resources when the driver is unloaded.

Signed-off-by: Thippeswamy Havalige <thippeswamy.havalige@amd.com>